Move relations code to higher order component
